Auto stash before merge of "main" and "origin/main"

+++ src/database/migrations/2021_03_05_221708_create_user_activities_table.php
@@ -13,20 +13,24 @@ class CreateUserActivitiesTable extends Migration
      */
     public function up()
     {
-        Schema::connection('mysql_second')->create('user_activities', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('type');
-            $table->unsignedInteger('parent_id')->nullable()->comment('Parent Activity ID');
-            $table->unsignedInteger('user_id')->nullable();
-            $table->unsignedInteger('object_id');
-            $table->string('object_type', 60);
-            $table->text('data')->nullable();
-            $table->timestamps();
-
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->foreign('parent_id')->references('id')->on('user_activities')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('user_activities')) {
+            Schema::create('user_activities', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('type');
+                $table->unsignedBigInteger('parent_id')->nullable()->comment('Parent Activity ID');
+                $table->unsignedBigInteger('user_id')->nullable();
+                $table->unsignedBigInteger('object_id');
+                $table->string('object_type', 60);
+                $table->text('data')->nullable();
+                $table->timestamps();
+    
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->foreign('parent_id')->references('id')->on('user_activities')->onDelete('cascade');
+            });
+        }
     }
+    
+    
 
     /**
      * Reverse the migrations.
